Major update
Short int changed to int, isBorder fixed - working properly now, end
game signalization added

+++ src/ai.cpp
@@ -41,8 +41,8 @@ AI::AI() {
 @brief Runs selected AI
 @return chosen index for Disk
 */
-short int AI::run(GameBoard * gb, Save * save) {
-	short int result = 0;
+int AI::run(GameBoard * gb, Save * save) {
+	int result = 0;
 	// select algorithm
 	switch (gb->getAIType()) {
 		case 1:
@@ -73,10 +73,10 @@ int AI::genRandom(int from, int to) {
 /*
 @brief Algorithm for choosing random position
 */
-short int AI::random(GameBoard * gb) {
-	vector <pair <short int, short int> > avail = gb->getAvail();
+int AI::random(GameBoard * gb) {
+	vector <pair <int, int> > avail = gb->getAvail();
 	// select random index from vector
-	short int index = genRandom(0, avail.size()-1);
+	int index = genRandom(0, avail.size()-1);
 	// wait 1.2 sec
 	this_thread::sleep_for(chrono::milliseconds(1200));
 
@@ -86,7 +86,7 @@ short int AI::random(GameBoard * gb) {
 /*
 @brief Evaluate position on Gameboard
 */
-double AI::evalPos(short int index, GameBoard * gb) {
+double AI::evalPos(int index, GameBoard * gb) {
 	int x = index / gb->size;
 	int y = index % gb->size;
 	int max = gb->size - 1;
@@ -109,12 +109,12 @@ double AI::evalPos(short int index, GameBoard * gb) {
 /*
 @brief Algorithm with simple heuristic 
 */
-short int AI::simpleH(GameBoard * gb, Save * save) {
+int AI::simpleH(GameBoard * gb, Save * save) {
 	bool loadRes;
 	size_t w_count = gb->getVec(WHITE).size();
 	double pos_val = 0.0;
 	int turn_val = 0;
-	vector <pair <short int, short int> > avail = gb->getAvail();
+	vector <pair <int, int> > avail = gb->getAvail();
 	// for storing value of available places
 	vector<double> value;
 	value.reserve(avail.size());
@@ -133,7 +133,7 @@ short int AI::simpleH(GameBoard * gb, Save * save) {
 	// max value in vector
 	auto max_value = max_element(value.begin(), value.end());
 	// index of max_value in vector
-	short int best_index = distance(value.begin(), max_value);
+	int best_index = distance(value.begin(), max_value);
 
 	// wait 1.2 sec
 	this_thread::sleep_for(chrono::milliseconds(1200));
